initial refactoring attempt
