+++ src/main/java/com/multicraft/MultiCraftCommandExecutor.java
@@ -1,5 +1,9 @@
 package com.multicraft;
 
+import com.multicraft.data.BlockRecord;
+import com.multicraft.data.BuildCommandData;
+import com.multicraft.data.PreviousBuildsData;
+import com.multicraft.exception.NoCommandHistoryException;
 import org.bukkit.Bukkit;
 import org.bukkit.GameMode;
 import org.bukkit.Location;
