+++ src/main/java/com/multicraft/exceptions/MaterialDoesNotExistException.java
@@ -1,0 +1,11 @@
+package com.multicraft.exception;
+
+public class MaterialDoesNotExistException extends Exception {
+
+    private static final long serialVersionUID = 0L;
+
+    public MaterialDoesNotExistException(String message) {
+        super(message);
+    }
+
+}
