+++ src/main/java/com/multicraft/exceptions/NoCommandHistoryException.java
@@ -1,12 +1,11 @@
-package com.multicraft;
+package com.multicraft.exception;
 
 public class NoCommandHistoryException extends Exception{
-	/**
-	 * 
-	 */
+
 	private static final long serialVersionUID = 1L;
 	
 	public NoCommandHistoryException(String message) {
 		super(message);
 	}
+
 }
