+++ src/main/java/com/multicraft/net/SpeechToTextServer.java
