+++ src/main/java/com/multicraft/util/JSONParser.java
@@ -1,13 +1,15 @@
-package com.multicraft;
+package com.multicraft.util;
 
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
 import org.json.simple.JSONValue;
 
 public class JSONParsing {
-	public static JSONObject JSONFromString(String str){
+
+	public static JSONObject JSONFromString(String str) {
 		Object obj = JSONValue.parse(str);
 		JSONArray array = (JSONArray) obj;
 		return (JSONObject) array.get(0);
 	}
+
 }
